+++ flask_security/cli.py
@@ -6,6 +6,7 @@
     Command Line Interface for managing accounts and roles.
 
     :copyright: (c) 2016 by CERN.
+    :copyright: (c) 2019 by J. Christopher Wagner
     :license: MIT, see LICENSE for more details.
 """
 
@@ -77,10 +78,17 @@ def users_create(identity, password, active):
 @roles.command("create")
 @click.argument("name")
 @click.option("-d", "--description", default=None)
+@click.option("-p", "--permissions")
 @with_appcontext
 @commit
 def roles_create(**kwargs):
     """Create a role."""
+
+    # For some reaosn Click puts arguments in kwargs - even if they weren't specified.
+    if "permissions" in kwargs and not kwargs["permissions"]:
+        del kwargs["permissions"]
+    if "permissions" in kwargs and not hasattr(_datastore.role_model, "permissions"):
+        raise click.UsageError("Role model does not support permissions")
     _datastore.create_role(**kwargs)
     click.secho('Role "%(name)s" created successfully.' % kwargs, fg="green")