+++ flask_security/models/__init__.py
@@ -1,0 +1,12 @@
+""""
+Copyright 2019 by J. Christopher Wagner (jwag). All rights reserved.
+:license: MIT, see LICENSE for more details.
+
+This packages contains OPTIONAL models for various ORMs/databases that can be used
+to quickly get the required DB models setup.
+
+These models have the fields for ALL features. This makes it easy for applications
+to add features w/o a DB migration (and modern DBs are pretty efficient at storing
+empty values!).
+
+"""